Cleanup and renames
